fix: misc fixes for codefresh, native build image and helm deploys (#33)
* use git org from app_setup config for codefresh steps

- for now using the config from app_setup.org. this still can be empty
with current setup
- when we build the onboarding flow, maybe make this required so this
config value can never be empty!

* fix: suffix for image_name should not be needed

no history on why we added this for now,  so to handle external ECR
setups, we dont want to add a suffix to the image name generated.

we need to test this change with codefresh and buildkit to make sure
this doesnt break anything

* fix: normalize git branch name for build and deploy job manifests

* fix: add version to public and org chart's nativeHelm deploy manifest

* fix: add --reset-values flag for nativeHelm seed row

* fix: duration for failed jobs in deployLogs table

* fix: duration for failed jobs in buildLogs api